requirements.txt
google-generativeai>=0.3.0
pypdf>=3.15.1
langchain>=0.0.267
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
streamlit>=1.25.0